use float instead of multiple possibly unsafe casts

+++ Firmware/Marlin-bugfix-2.0.9.3.x/Marlin/src/feature/hx_711.h
@@ -68,7 +68,7 @@ public:
      *
      * @param thrIn
      */
-    void setThreshold(int32_t thrIn) { _th_weigth = (float)thrIn; };
+    void setThreshold(float thrIn) { _th_weigth = thrIn; };
 
     /**
      * @brief Set the HX711 channel/mode.
